requirements.txt
boto3
jpype1
aspose-cells
